resume-builder/pyproject.toml
[tool.poetry]
name = "resume-builder"
version = "0.1.0"
description = ""
authors = ["nalin <74811063+lordspline@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.49.1"
scrapybara = "^2.0.7"
python-dotenv = "^1.0.1"
httpx = "0.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
